pyproject.toml
[tool.poetry]
name = "2-langserve-basic-poetry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.2.2"}
pydantic = ">=2.7.4,<3.0.0"  
langchain-openai = "^0.1.19"
python-decouple = "^3.8"
langsmith = "^0.1.96"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
